requirements.txt
pygame==1.9.6
PyOpenGL==3.1.5
pillow==7.1.2
pywavefront==1.3.2
numpy==1.18.3
PyGLM==1.2.0
screeninfo==0.6.5
